+++ src/main/kotlin/de/achimonline/ansible_vault_integration/intention/AnsibleVaultIdentityPopup.kt
@@ -4,7 +4,7 @@ import com.intellij.icons.AllIcons
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.util.ui.JBUI
-import de.achimonline.ansible_vault_integration.bundle.AnsibleVaultIntegrationBundle
+import de.achimonline.ansible_vault_integration.bundle.AnsibleVaultIntegrationBundle.message
 import de.achimonline.ansible_vault_integration.config.VaultIdentity
 import java.awt.Component
 import java.util.function.Consumer
@@ -25,8 +25,8 @@ class AnsibleVaultIdentityPopup(
 
     private val popup = JBPopupFactory.getInstance().createPopupChooserBuilder(vaultIdentities)
         .setSelectionMode(ListSelectionModel.SINGLE_SELECTION)
-        .setAccessibleName(AnsibleVaultIntegrationBundle.message("settings.vault_identity.accessible_name"))
-        .setTitle(AnsibleVaultIntegrationBundle.message("settings.vault_identity.title"))
+        .setAccessibleName(message("settings.vault_identity.accessible_name"))
+        .setTitle(message("settings.vault_identity.title"))
         .setMovable(false)
         .setResizable(false)
         .setRequestFocus(true)